feat: Start using getPlaceById
(130, 10) <!-- TODO (9) Add the "Powered by Google" logo and a link to Google's privacy policy here -->
MainActivity.java
(81, 12) // TODO (3) Modify the Adapter to take a PlaceBuffer in the constructor
(99, 8) // TODO (1) Implement a method called refreshPlacesData that:
(140, 7) //TODO (8) Set the getPlaceById callBack so that onResult calls the Adapter's swapPlaces with the result
(142, 7) //TODO (2) call refreshPlacesData in GoogleApiClient's onConnected and in the Add New Place button click event
PlaceListAdapter.java
(40, 12) // TODO (4) Take a PlaceBuffer as an input and store it as a local private member mPlaces
(68, 12) // TODO (6) Implement onBindViewHolder to set the view holder's Name and Address text fields
(80, 7) //TODO (7) Implement a public method swapPlaces that replaces the current mPlaces PlaceBuffer with a new one
(99, 12) // TODO (5) Update getItemCount to return mPlaces's item count

Get place by ID
For this exercise, we've seen how to use getPlaceById() method to retrieve the
Places details from the server and display them in the recycler view. We've also
seen how important it is to attribute any Google Places API data to google using
the "powered by Google" logo.