feat: Set Silent Mode when Enter/Exiting a Geofence
Now the last piece of our puzzle we need to actually set the device to silent
when a Geofence triggers an entry transition and change back to normal when an
exit transition is triggered!

(100, 14) <!-- TODO (1) Add a LinearLayout with the permissions CheckBox and UI -->
layout (1 item in 1 file)
activity_main.xml
(100, 14) <!-- TODO (1) Add a LinearLayout with the permissions CheckBox and UI -->
com.example.android.shushme (5 items in 2 files)
GeofenceBroadcastReceiver.java
(48, 12) // TODO (4) Use GeofencingEvent.fromIntent to retrieve the GeofencingEvent that caused the transition
(53, 12) // TODO (5) Call getGeofenceTransition to get the transition type and use AudioManager to set the
(72, 12) // TODO (6) Show a notification to alert the user that the ringer mode has changed.
MainActivity.java
(256, 11) //TODO (3) Initialize ringer permissions checkbox
(275, 8) // TODO (2) Implement onRingerPermissionsClicked to launch ACTION_NOTIFICATION_POLICY_ACCESS_SETTINGS