fix: Add Carlos' fix for the Too Long ID problem
This is the answer you got to your questions from Carlos.

Question::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::

Hey Carlos, how are you? I have some questions about the ShushMe app in
Advanced Android Apps, the final project(Silent Mode). In my actual app and
the solution, when I click Ringer permissions after enabling all the things on
the main page, I get the Do Not Disturb page with 4 choices with checkboxes.
None of the choices is the app I just created. Is that what is supposed to
happen? I click 1 or more of the choices on this page and return. Upon return
to the previous page the box I just checked is unchecked. I am not sure if this
is what is supposed to happen either.

Secondly, I sometimes get (java.lang.IllegalArgumentException: requestId is null
or too long:) even when I am using the Solution for the last project. This is
when I type in an address. The ID is not null, but I have nothing to do with it
being long.

Lastly, I know this was somewhat covered in the videos, but how do I test the
Ringer? I put landmarks in my MainActivity, then I go to the, “…”, in emulator
and put in some GPS coordinates, but nothing happens. I am a little confused.

Here is the course solution:
https://github.com/udacity/AdvancedAndroid_Shushme/tree/T0X.05-Solution-SilentMode 

Too long, error(This is the PlaceId I was given):
java.lang.IllegalArgumentException: requestId is null or too long: EjMyMTAxIE5XIDIzcmQgU3QsIE9rbGFob21hIENpdHksIE9rbGFob21hIDczMTA3LCBVU0EiMRIvChQKEglDd7iFlxCyhxFVPGriC8o6txC1ECoUChIJRYBS0HIQsocRKgbhm9yiXm0

Thanks, a lot!

Answer:::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::

To be able to solve the issue with the Ringer permissions, please add the
following line on your manifest and the ShushMe app will appear on the "Do
not Disturb Access" setting:

<uses-permission android:name="android.permission.ACCESS_NOTIFICATION_POLICY" />

I tested the solution and everything works well on my API 27 Emulator, I'm
not getting that error with the requestId.

To test the ringer, you need to check if the GPS that you are using is within
50 meters of the address added, that is a bit complicated to know if you search
by area when you add a place and not by the exact address. Another option is to
increase the radius on the Geofencing class.

Hope it helps :)